Form closing cleanups
Cleaned up how the application handles forms closing.
The MainMenu form is set to hide itself instead of closing because the
dispose method closes all forms when closing the startup form, I haven't
figured out how to fix that.

The Game form has a exit bool set to true by default. When the main menu
button is pressed, the exit bool is set to false. During the form's
close event Application.Exit() will be called only if the exit bool is
true.